services/renderer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
manim==0.20.1
google-cloud-texttospeech
google-cloud-storage==2.18.0
